refactor test to cut down some complexity
