+++ __tests__/app/(dashboard)/add/page.test.tsx
@@ -1,20 +1,20 @@
 import '@testing-library/jest-dom'
 import Page from "@/app/(dashboard)/dashboard/add/page"
-import {render, screen} from "@testing-library/react"
+import {render} from "@testing-library/react"
 
 jest.mock('../../../../src/components/AddFriendButton', () => ({
     __esModule: true,
     default: () => <div data-testid="friend-button"/>
-}))
+}));
 
 describe('Add Page tests',()=>{
    test('Document should instruct you to add a friend',  ()=>{
-       const expected = 'Add a Friend'
-       render(<Page/>)
-       expect(screen.getByText(expected)).toBeInTheDocument()
-    })
+       const {getByText} = render(<Page/>);
+       expect(getByText('Add a Friend')).toBeInTheDocument();
+    });
+
     test('needs to contain an add-friends button',()=>{
-        const {getByTestId} = render(<Page/>)
+        const {getByTestId} = render(<Page/>);
         expect(getByTestId('friend-button')).toBeInTheDocument()
-    })
-})
+    });
+});
